+++ src/main/java/shop/tryit/domain/answer/Answer.java
@@ -13,7 +13,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import shop.tryit.domain.common.BaseTimeEntity;
 import shop.tryit.domain.member.Member;
-import shop.tryit.domain.question.Question;
+import shop.tryit.domain.question.entity.Question;
 
 @Entity
 @Getter
