+++ src/main/java/shop/tryit/domain/answer/AnswerRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import shop.tryit.domain.question.Question;
+import shop.tryit.domain.question.entity.Question;
 
 public interface AnswerRepository {
