+++ src/main/java/shop/tryit/domain/answer/service/AnswerService.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.tryit.domain.answer.Answer;
 import shop.tryit.domain.answer.AnswerRepository;
-import shop.tryit.domain.question.Question;
+import shop.tryit.domain.question.entity.Question;
 import shop.tryit.domain.question.service.QuestionService;
 
 @Service
