+++ src/main/java/shop/tryit/repository/answer/AnswerJpaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import shop.tryit.domain.answer.Answer;
-import shop.tryit.domain.question.Question;
+import shop.tryit.domain.question.entity.Question;
 
 public interface AnswerJpaRepository extends JpaRepository<Answer, Long> {
