+++ src/main/java/shop/tryit/repository/question/QuestionJpaRepository.java
@@ -1,7 +1,8 @@
 package shop.tryit.repository.question;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.tryit.domain.question.Question;
+import shop.tryit.domain.question.entity.Question;
 
 public interface QuestionJpaRepository extends JpaRepository<Question, Long>, QuestionCustom {
+
 }
